logout problem securt

+++ lib/screens/children/blocs/children_management_bloc/children_management_bloc.dart
@@ -1,5 +1,6 @@
 import 'package:bloc/bloc.dart';
 import 'package:child_repository/child_repository.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:perfect_childcare/screens/children/blocs/children_management_bloc/children_management_event.dart';
 import 'package:perfect_childcare/screens/children/blocs/children_management_bloc/children_management_state.dart';
 import 'package:user_repository/user_repository.dart';
@@ -19,9 +20,15 @@ class ChildrenManagementBloc
   }
 
   Future<void> _onAddChild(
-      AddChildEvent event, Emitter<ChildrenManagementState> emit) async {
+    AddChildEvent event,
+    Emitter<ChildrenManagementState> emit,
+  ) async {
     emit(ChildrenManagementLoading());
     try {
+      if (FirebaseAuth.instance.currentUser == null) {
+        emit(const ChildrenManagementFailure('User not logged in'));
+        return;
+      }
       await childRepository.addChild(event.child);
       // Update user's children list
       await userRepository.connectChildToUser(event.userId, event.child.id);
@@ -45,6 +52,10 @@ class ChildrenManagementBloc
       RemoveChildEvent event, Emitter<ChildrenManagementState> emit) async {
     emit(ChildrenManagementLoading());
     try {
+      if (FirebaseAuth.instance.currentUser == null) {
+        emit(const ChildrenManagementFailure('User not logged in'));
+        return;
+      }
       await userRepository.disconnectChildFromUser(event.childId);
       await childRepository.removeChild(event.childId);
 
@@ -58,6 +69,10 @@ class ChildrenManagementBloc
       UpdateChildEvent event, Emitter<ChildrenManagementState> emit) async {
     emit(ChildrenManagementLoading());
     try {
+      if (FirebaseAuth.instance.currentUser == null) {
+        emit(const ChildrenManagementFailure('User not logged in'));
+        return;
+      }
       await childRepository.updateChild(event.child);
       emit(ChildrenManagementSuccess());
     } catch (e) {
